Create SourcesLayer component
